jfc-web-app/requirements.txt
absl-py==2.2.2
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiohttp-cors==0.7.0
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.9.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.1.0
autocommand==2.2.2
Automat==24.8.1
babel==2.17.0
backports.tarfile==1.2.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
bloodhound==1.8.0
boltons==21.0.0
bracex==2.5.post1
build==1.2.2.post1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-option-group==0.5.7
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
contourpy==1.3.1
cryptography==42.0.8
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
deepdiff==5.8.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
Django==5.2.5
djangorestframework==3.16.1
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.2.0
face==24.0.0
fastapi==0.115.8
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.5.1
future==1.0.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
glom==22.1.0
gns3-gui==2.2.53
gns3-server==2.2.53
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
hf-xet==1.1.5
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.33.2
hyperlink==21.0.0
idna==3.10
impacket==0.12.0
importlib_metadata==7.1.0
incremental==24.7.2
inflect==7.3.1
installer==0.7.0
ipykernel==6.29.5
ipympl==0.9.6
ipython==8.32.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
javalang==0.13.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
joserfc==1.0.3
json5==0.10.0
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.10.0
keras-metrics==1.1.0
keras-tuner==1.4.7
kiwisolver==1.4.8
kt-legacy==1.0.5
ldap3==2.9.1
ldapdomaindump==0.9.4
libclang==18.1.1
loguru==0.6.0
Mako==1.3.9
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.1
mitm6==0.3.0
ml-dtypes==0.4.1
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.1.0
Naked==0.1.32
namex==0.0.9
narwhals==1.41.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
neo4j==5.28.0
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==2.6.3
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.0.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
opt_einsum==3.4.0
optree==0.15.0
ordered-set==4.1.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
peewee==3.18.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.1.2
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pydot==1.4.1
Pygments==2.19.1
pyOpenSSL==24.0.0
pyparsing==3.2.1
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.3
rpds-py==0.22.3
ruamel.yaml==0.18.12
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scapy==2.6.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semgrep==1.123.0
Send2Trash==1.8.3
sentry-sdk==2.22.0
service-identity==24.2.0
shellescape==3.8.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.45.3
sympy==1.14.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.21.2
tomli==2.0.1
torch==2.7.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.0
tree_sitter==0.20.1
triton==3.3.1
truststore==0.10.1
Twisted==24.11.0
typeguard==4.3.0
typer==0.4.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
wandb==0.21.0
watchdog==6.0.0
wcmatch==8.5.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
yarl==1.18.3
zipp==3.19.2
zope.interface==7.2
zstandard==0.23.0
